Fixed destIndex value passed to accept callback when dragging below element
I've had an issue for a while where the accept callback gets a destIndex of 0 regardless of whether I was dragging to the first position or the second within a node. I traced down the call to the accept callback and found that if I increment the index returned if targetBefore is false then I received the correct index within my accept callback. I noticed that the library is already doing something similar directly following the call to accept when calling moveTo on the dragInfo..